Rename `TakerNegotiator` to `Negotiator` internally, still export as `TakerNegotiator`
